Move ChatMessageLayout type to ChatItem level
